Accounts
Accounts are somewhat functional. There is login and register support
via session storage and an array of users. Any registered users are lost
upon closing all tabs. Users are logged out upon closing all tabs.
Potential improvements to this design include pushing user data onto a
database such as SQL. My Account page is still untouched and planned to
be completed this week. The storage of the data associated with the page
is best suited for a database.